fix: Use large data types for files sizes where appropriate.
Switch data types to use int64_t instead of `size_t`, which is only 32
bits on 32-bit platforms, in places where we expect to handle file
sizes.

Also enable error messages when we try to auto-convert from one type
to another in a lossy fashion.

Changelog: Fix possible integer overflow when dealing with large files
on 32-bit platforms.

Ticket: MEN-7613

+++ src/artifact/v3/payload/payload.hpp
@@ -54,7 +54,7 @@ public:
 	string Name() {
 		return this->entry_->Name();
 	}
-	size_t Size() {
+	int64_t Size() {
 		return this->entry_->Size();
 	}
